Split tokens into multiple files.
