+++ experiments/react-native-uifabric/src/components/PersonaCoin/PersonaCoin.settings.ts
@@ -11,10 +11,7 @@ export const settings: IComposeSettings<IPersonaCoinType> = [
       horizontalIconAlignment: 'end',
       verticalIconAlignment: 'end',
       color: 'white', // initials is always 'white', unless overriden by token
-      // initialsSize: defaultSizeConfig.initialsSize,
-      // coinSize: defaultSizeConfig.physicalSize,
-      // iconSize: defaultSizeConfig.iconSize,
-      backgroundColor: convertCoinColor('lightBlue')
+      backgroundColor: convertCoinColor(defaultColor)
     }
   },
   personaCoinName
