+++ experiments/react-native-uifabric/src/components/PersonaCoin/PersonaCoin.tsx
@@ -1,5 +1,5 @@
 /** @jsx withSlots */
-import { Image, View, ImageURISource, Text } from 'react-native';
+import { Image, View, Text } from 'react-native';
 import {
   IPersonaCoinProps,
   IPersonaCoinType,
@@ -13,38 +13,39 @@ import { filterViewProps, filterImageProps } from '../../utilities/RenderHelpers
 import { settings } from './PersonaCoin.settings';
 import { ISlots, withSlots, IRenderData } from '@uifabricshared/foundation-composable';
 import { mergeSettings } from '@uifabricshared/foundation-settings';
-import {  getPresenceIconSource } from './PersonaCoin.helpers';
-import { buildRootStyles, buildPhotoStyles, buildInitialsStyles, buildInitialsBackgroundStyles, buildIconStyles } from './PersonaCoin.tokens';
+import { getPresenceIconSource } from './PersonaCoin.helpers';
 import { foregroundColorTokens } from '../../tokens';
+import { buildRootStyles } from './PersonaCoin.tokens.root';
+import { buildInitialsStyles } from './PersonaCoin.tokens.initials';
+import { buildInitialsBackgroundStyles } from './PersonaCoin.tokens.initialsBackground';
+import { buildPhotoStyles } from './PersonaCoin.tokens.photo';
+import { buildIconStyles } from './PersonaCoin.tokens.icon';
 
 function usePrepareForProps(
   props: IPersonaCoinProps,
   useStyling: IUseComposeStyling<IPersonaCoinType>
 ): IRenderData<IPersonaCoinSlotProps, IPersonaCoinState> {
-  const { imageUrl, imageDescription, initials,  presence, ...rest } = props;
+  const { imageUrl, imageDescription, initials, presence, ...rest } = props;
 
-  let personaPhotoSource: ImageURISource | undefined = undefined;
-  if (imageUrl) {
-    personaPhotoSource = {
-      uri: imageUrl
-    };
-  }
+  const personaPhotoSource =
+    imageUrl === undefined
+      ? undefined
+      : {
+          uri: imageUrl
+        };
 
-  let iconSource: ImageURISource | undefined = undefined;
-  if (presence) {
-    iconSource = getPresenceIconSource(presence);
-  }
+  const iconSource = presence === undefined ? undefined : getPresenceIconSource(presence);
 
   return {
     slotProps: mergeSettings<IPersonaCoinType['slotProps']>(useStyling(props), {
       root: { ...rest },
       initials: {
-        children: initials,
+        children: initials
       },
       photo: {
         accessibilityLabel: imageDescription,
-        resizeMode: 'cover',
-      },
+        resizeMode: 'cover'
+      }
     }),
     state: {
       iconSource,
@@ -69,7 +70,7 @@ const render = (Slots: ISlots<IPersonaCoinSlotProps>, renderData: IPersonaCoinRe
           <Slots.initials />
         </Slots.initialsBackground>
       )}
-      {!!iconSource && <Slots.icon source={iconSource} />}
+      {!!iconSource && !!iconSource.uri && <Slots.icon source={iconSource} />}
     </Slots.root>
   );
 };
@@ -103,6 +104,6 @@ export const PersonaCoin = compose<IPersonaCoinType>({
     initials: [foregroundColorTokens, buildInitialsStyles],
     initialsBackground: [buildInitialsBackgroundStyles],
     photo: [buildPhotoStyles],
-    icon: [buildIconStyles],
+    icon: [buildIconStyles]
   }
 });
