+++ experiments/tester/src/RNTester/TestComponents/PersonaCoin/StandardUsage.tsx
@@ -35,6 +35,10 @@ export const StandardUsage: React.FunctionComponent<{}> = () => {
   const [coinColor, setCoinColor] = React.useState<PersonaCoinColor>('gold');
   const [presence, setPresence] = React.useState<PersonaPresence>('online');
 
+  const onSizeChange = React.useCallback(value => setImageSize(value), []);
+  const onColorChange = React.useCallback(value => setCoinColor(value), []);
+  const onPresenceChange = React.useCallback(value => setPresence(value), []);
+
   return (
     <View style={styles.root}>
       {/* settings */}
@@ -44,19 +48,19 @@ export const StandardUsage: React.FunctionComponent<{}> = () => {
           <Switch value={showImage} onValueChange={setShowImage} />
         </View>
 
-        <Picker prompt="Size" style={styles.header} selectedValue={imageSize} onValueChange={setImageSize}>
+        <Picker prompt="Size" style={styles.header} selectedValue={imageSize} onValueChange={onSizeChange}>
           {allSizes.map((size, index) => (
             <Picker.Item label={size} key={index} value={size} />
           ))}
         </Picker>
 
-        <Picker prompt="Coin color" style={styles.header} selectedValue={coinColor} onValueChange={setCoinColor}>
+        <Picker prompt="Coin color" style={styles.header} selectedValue={coinColor} onValueChange={onColorChange}>
           {allColors.map((color, index) => (
             <Picker.Item label={color} key={index} value={color} />
           ))}
         </Picker>
 
-        <Picker prompt="Presence status" style={styles.header} selectedValue={presence} onValueChange={setPresence}>
+        <Picker prompt="Presence status" style={styles.header} selectedValue={presence} onValueChange={onPresenceChange}>
           {allPresences.map((presence, index) => (
             <Picker.Item label={presence} key={index} value={presence} />
           ))}
